+++ src/main/java/lotto/domain/Lotto.java
@@ -55,21 +55,4 @@ public abstract class Lotto {
     }
 
     public abstract boolean isAuto();
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) {
-            return true;
-        }
-        if (!(o instanceof Lotto)) {
-            return false;
-        }
-        Lotto lotto = (Lotto) o;
-        return lottoNumbers.equals(lotto.lottoNumbers);
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(lottoNumbers);
-    }
 }
