+++ src/main/java/lotto/util/RandomGenerator.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.util;
 
 import java.util.Random;
 import java.util.concurrent.ThreadLocalRandom;
